refactor(agent): split daemon.rs into daemon/{manager,policy_cache,audit_queue,...}
Convert the single 3,339-line daemon.rs into a daemon/ submodule directory
where each production file is under 600 lines. Public re-exports preserve
the previous module surface so call sites in main.rs, api_server.rs,
policy.rs, posture_commands.rs, tray.rs, updater.rs, and edr/dto.rs are
unaffected.

Layout:
  * state.rs              - DaemonState/DaemonConfig/DaemonStatus/HealthResponse
  * manager.rs            - DaemonManager lifecycle + health monitor
  * spawn.rs              - child spawn/terminate/backoff helpers
  * ready_probe.rs        - /health probing and ready-wait loops
  * runtime_config.rs     - hushd.runtime.yaml writer (uses hushd-config crate)
  * runtime_keypair.rs    - per-port signing/enrollment key materialization
  * policy_cache.rs       - last-known-good policy bundle cache
  * audit_queue.rs        - persistent audit-event outbox with chunked flush
  * audit_queue_tests.rs  - audit_queue tests (kept in dedicated file to keep
                            audit_queue.rs production code under 600 lines)
  * binary_discovery.rs   - managed hushd binary path/seed/lookup

Tests live alongside the module they exercise so each suite can reach the
private internals it needs without re-exporting them.
